+++ src/EditGuesser.test.js
@@ -1,10 +1,15 @@
 import Api from '@api-platform/api-doc-parser/lib/Api';
 import Field from '@api-platform/api-doc-parser/lib/Field';
 import Resource from '@api-platform/api-doc-parser/lib/Resource';
-import {ArrayInput, SimpleFormIterator, TextInput} from 'react-admin';
+import {
+  ArrayInput,
+  DisabledInput,
+  SimpleFormIterator,
+  TextInput,
+} from 'react-admin';
 import {shallow} from 'enzyme';
 import React from 'react';
-import Create from './Create';
+import EditGuesser from './EditGuesser';
 import inputFactory from './inputFactory';
 
 const entrypoint = 'http://entrypoint';
@@ -45,7 +50,7 @@ const resourceData = {
   ],
 };
 
-describe('<Create />', () => {
+describe('<EditGuesser />', () => {
   test('without overrides', () => {
     const defaultInputFactory = jest.fn(inputFactory);
 
@@ -61,7 +66,7 @@ describe('<Create />', () => {
     });
 
     const render = shallow(
-      <Create
+      <EditGuesser
         options={{
           api,
           inputFactory: defaultInputFactory,
@@ -75,7 +80,46 @@ describe('<Create />', () => {
 
     expect(defaultInputFactory).toHaveBeenCalledTimes(3);
     expect(render.find(ArrayInput).length).toEqual(1);
+    expect(render.find(DisabledInput).length).toEqual(1);
+    expect(render.find(DisabledInput).get(0).props.source).toEqual('id');
     expect(render.find(SimpleFormIterator).length).toEqual(1);
+    expect(render.find(TextInput).length).toEqual(3);
+    expect(render.find(ArrayInput).get(0).props.source).toEqual('fieldD');
+    expect(render.find(TextInput).get(0).props.source).toEqual('fieldA');
+    expect(render.find(TextInput).get(1).props.source).toEqual('fieldB');
+  });
+
+  test('without default identifier input', () => {
+    const defaultInputFactory = jest.fn(inputFactory);
+
+    const resource = new Resource(resourceData.name, resourceData.url, {
+      ...resourceData,
+      editProps: {
+        addIdInput: false,
+      },
+    });
+
+    const api = new Api(apiData.entrypoint, {
+      ...apiData,
+      resources: [resource],
+    });
+
+    const render = shallow(
+      <EditGuesser
+        options={{
+          api,
+          inputFactory: defaultInputFactory,
+          resource,
+        }}
+        location={{}}
+        match={{}}
+        resource=""
+      />,
+    );
+
+    expect(defaultInputFactory).toHaveBeenCalledTimes(3);
+    expect(render.find(ArrayInput).length).toEqual(1);
+    expect(render.find(DisabledInput).length).toEqual(0);
     expect(render.find(TextInput).length).toEqual(3);
     expect(render.find(ArrayInput).get(0).props.source).toEqual('fieldD');
     expect(render.find(TextInput).get(0).props.source).toEqual('fieldA');
@@ -88,7 +132,7 @@ describe('<Create />', () => {
 
     const resource = new Resource(resourceData.name, resourceData.url, {
       ...resourceData,
-      createProps: {
+      editProps: {
         options: {
           inputFactory: customInputFactory,
         },
@@ -101,7 +145,7 @@ describe('<Create />', () => {
     });
 
     const render = shallow(
-      <Create
+      <EditGuesser
         options={{
           api,
           inputFactory: defaultInputFactory,
@@ -116,6 +160,8 @@ describe('<Create />', () => {
     expect(customInputFactory).toHaveBeenCalledTimes(3);
     expect(defaultInputFactory).toHaveBeenCalledTimes(0);
     expect(render.find(ArrayInput).length).toEqual(1);
+    expect(render.find(DisabledInput).length).toEqual(1);
+    expect(render.find(DisabledInput).get(0).props.source).toEqual('id');
     expect(render.find(SimpleFormIterator).length).toEqual(1);
     expect(render.find(TextInput).length).toEqual(3);
     expect(render.find(ArrayInput).get(0).props.source).toEqual('fieldD');
@@ -128,7 +174,7 @@ describe('<Create />', () => {
 
     const resource = new Resource(resourceData.name, resourceData.url, {
       ...resourceData,
-      createFields: [
+      editFields: [
         new Field('fieldC', {
           id: 'http://schema.org/fieldC',
           range: 'http://www.w3.org/2001/XMLSchema#string',
@@ -150,7 +196,7 @@ describe('<Create />', () => {
     });
 
     const render = shallow(
-      <Create
+      <EditGuesser
         options={{
           api,
           inputFactory: defaultInputFactory,
@@ -164,9 +210,59 @@ describe('<Create />', () => {
 
     expect(defaultInputFactory).toHaveBeenCalledTimes(2);
     expect(render.find(ArrayInput).length).toEqual(1);
+    expect(render.find(DisabledInput).length).toEqual(1);
+    expect(render.find(DisabledInput).get(0).props.source).toEqual('id');
     expect(render.find(SimpleFormIterator).length).toEqual(1);
     expect(render.find(TextInput).length).toEqual(2);
     expect(render.find(ArrayInput).get(0).props.source).toEqual('fieldE');
     expect(render.find(TextInput).get(0).props.source).toEqual('fieldC');
+  });
+
+  test('with writable identifier', () => {
+    const defaultInputFactory = jest.fn(inputFactory);
+
+    const resource = new Resource(resourceData.name, resourceData.url, {
+      ...resourceData,
+      editProps: {
+        addIdInput: false,
+      },
+      writableFields: [
+        ...resourceData.writableFields,
+        new Field('id', {
+          id: 'http://schema.org/identifier',
+          range: 'http://www.w3.org/2001/XMLSchema#string',
+          reference: null,
+          required: true,
+        }),
+      ],
+    });
+
+    const api = new Api(apiData.entrypoint, {
+      ...apiData,
+      resources: [resource],
+    });
+
+    const render = shallow(
+      <EditGuesser
+        options={{
+          api,
+          inputFactory: defaultInputFactory,
+          resource,
+        }}
+        location={{}}
+        match={{}}
+        resource=""
+      />,
+    );
+
+    expect(defaultInputFactory).toHaveBeenCalledTimes(4);
+    expect(render.find(ArrayInput).length).toEqual(1);
+    expect(render.find(DisabledInput).length).toEqual(0);
+    expect(render.find(TextInput).length).toEqual(4);
+    expect(render.find(ArrayInput).get(0).props.source).toEqual('fieldD');
+    expect(render.find(TextInput).get(0).props.source).toEqual('fieldA');
+    expect(render.find(TextInput).get(1).props.source).toEqual('fieldB');
+    expect(render.find(TextInput).get(2).props.source).toEqual(undefined);
+    expect(render.find(TextInput).get(3).props.source).toEqual('id');
   });
 });
